+++ еests/_ests.swift
@@ -1,0 +1,16 @@
+//
+//  _ests.swift
+//  еests
+//
+//  Created by Mikhail Korzh on 15.09.2026.
+//
+
+import Testing
+
+struct _ests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}